requirements.txt
flask==3.0.0
gunicorn==21.2.0
retell-sdk==4.27.0
twilio==9.5.2
python-dotenv==1.0.0 